Tuan5

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -9,6 +9,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Models\VendorProductImageGallery;
 use App\Traits\ImageUploadTrait;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -19,6 +20,10 @@ class VendorProductImageGalleryController extends Controller
     public function index(Request $request, VendorProductImageGalleryDataTable $dataTable)
     {
         $product= Product::findOrFail($request->product);
+        /** Check product vendor */
+        if($product->vendor_id!= Auth::user()->vendor->id){
+            abort(404);
+        }
         return $dataTable->render('vendor.product.image-gallery.index',compact('product'));
     }
 
@@ -81,6 +86,10 @@ class VendorProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $productImage= ProductImageGallery::findOrFail($id);
+        /** Check Product vendor */
+        if($productImage->product->vendor_id!=Auth::user()->vendor->id){
+            abort(404);
+        }
         $this->deleteImage($productImage->image);
         $productImage->delete();
